Made Address, Duration, HumanName and Ratio version agnostic
