+++ src/Hl7.Fhir.Base/Model/Ratio.cs
